requirements.txt
# Core dependencies
llama-index==0.12.33
llama-index-embeddings-fastembed==0.3.1
llama-index-llms-ollama==0.1.1
llama-index-vector-stores-qdrant==0.1.3
qdrant-client==1.14.2
fastembed==0.6.1
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.26.4

# UI and web framework
gradio==4.44.1

# AWS integration
boto3==1.34.0
warrant==0.6.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Utilities
python-dotenv==1.1.0
